Exp/reimpl trapdoor (#63)
obfuscation time increase like 2.5x(20m for dummy) but memory bound like
half, evaluation time decreased even

---------
