+++ Cargo.toml
@@ -10,7 +10,7 @@ default = ["test"]
 test = []
 
 [dependencies]
-openfhe = { git = "https://github.com/MachinaIO/openfhe-rs.git" }
+openfhe = { git = "https://github.com/MachinaIO/openfhe-rs.git", branch = "exp/reimpl_trapdoor" }
 digest = "0.10"
 num-bigint = { version = "0.4", default-features = false }
 num-traits = "0.2"
@@ -28,9 +28,9 @@ memmap2 = "0.9.5"
 tempfile = "3.19.1"
 sysinfo = "0.34.1"
 once_cell = "1.21.1"
+rand_distr = "0.5.1"
 dashmap = "6.1.0"
 # for now we put in test
-
 
 [dev-dependencies]
 keccak-asm = { version = "0.1.4" }
