+++ src/io/eval.rs
@@ -83,27 +83,12 @@ where
         let log_q = params.as_ref().modulus_bits();
         let dim = params.as_ref().ring_dimension() as usize;
         for (idx, input) in inputs.iter().enumerate() {
-            let m_preimage_paths = if *input {
-                &self.m_preimages_paths[idx][1]
-            } else {
-                &self.m_preimages_paths[idx][0]
-            };
-            let m = ST::preimage_from_fs(params.as_ref(), m_preimage_paths);
-            let q = ps[idx].clone() * &m;
-            let n_preimage_paths = if *input {
-                &self.n_preimages_paths[idx][1]
-            } else {
-                &self.n_preimages_paths[idx][0]
-            };
-            let n = ST::preimage_from_fs(params.as_ref(), n_preimage_paths);
-            let p = q.clone() * &n;
-            let k_preimage_paths = if *input {
-                &self.k_preimages_paths[idx][1]
-            } else {
-                &self.k_preimages_paths[idx][0]
-            };
-            let k = ST::preimage_from_fs(params.as_ref(), k_preimage_paths);
-            let v = q.clone() * &k;
+            let m = if *input { &self.m_preimages[idx][1] } else { &self.m_preimages[idx][0] };
+            let q = ps[idx].clone() * m;
+            let n = if *input { &self.n_preimages[idx][1] } else { &self.n_preimages[idx][0] };
+            let p = q.clone() * n;
+            let k = if *input { &self.k_preimages[idx][1] } else { &self.k_preimages[idx][0] };
+            let v = q.clone() * k;
             let new_encode_vec = {
                 let t = if *input { &public_data.rgs[1] } else { &public_data.rgs[0] };
                 let encode_vec = encodings[idx][0].concat_vector(&encodings[idx][1..]);
@@ -208,8 +193,8 @@ where
             .collect_vec();
         let output_encodings_vec =
             output_encoding_ints[0].concat_vector(&output_encoding_ints[1..]);
-        let final_preimage = ST::preimage_from_fs(params.as_ref(), &self.final_preimage_path);
-        let final_v = ps.last().unwrap().clone() * &final_preimage;
+        let final_preimage = &self.final_preimage;
+        let final_v = ps.last().unwrap().clone() * final_preimage;
         let z = output_encodings_vec.clone() - final_v.clone();
         debug_assert_eq!(z.size(), (1, packed_output_size));
         #[cfg(feature = "test")]
