+++ src/io/obf.rs
@@ -123,10 +123,10 @@ where
     };
     log_mem("Computed u_0, u_1, u_star");
 
-    let (mut m_preimages_paths, mut n_preimages_paths, mut k_preimages_paths) = (
-        vec![vec![vec![]; 2]; obf_params.input_size],
-        vec![vec![vec![]; 2]; obf_params.input_size],
-        vec![vec![vec![]; 2]; obf_params.input_size],
+    let (mut m_preimages, mut n_preimages, mut k_preimages) = (
+        vec![Vec::with_capacity(2); obf_params.input_size],
+        vec![Vec::with_capacity(2); obf_params.input_size],
+        vec![Vec::with_capacity(2); obf_params.input_size],
     );
 
     #[cfg(feature = "test")]
@@ -168,32 +168,30 @@ where
             {
                 bs[idx + 1][bit] = b_bit_idx.clone();
             }
-            let m_preimage_bit_id = format!("m_preimage_{}_{}", idx, bit);
-            let n_preimage_bit_id = format!("n_preimage_{}_{}", idx, bit);
-            let k_preimage_bit_id = format!("k_preimage_{}_{}", idx, bit);
+            // let m_preimage_bit_id = format!("m_preimage_{}_{}", idx, bit);
+            // let n_preimage_bit_id = format!("n_preimage_{}_{}", idx, bit);
+            // let k_preimage_bit_id = format!("k_preimage_{}_{}", idx, bit);
 
-            let m_preimage_bit_path = sampler_trapdoor.preimage_to_fs(
+            let m_preimage_bit = sampler_trapdoor.preimage(
                 &params,
                 &b_star_trapdoor_cur,
                 &b_star_cur,
                 &(u_bits[bit].clone() * &b_bit_idx),
-                &m_preimage_bit_id,
             );
 
             log_mem("Computed m_preimage_bit");
 
-            m_preimages_paths[idx][bit] = m_preimage_bit_path;
+            m_preimages[idx].push(m_preimage_bit);
 
-            let n_preimage_bit_path = sampler_trapdoor.preimage_to_fs(
+            let n_preimage_bit = sampler_trapdoor.preimage(
                 &params,
                 &b_bit_trapdoor_idx,
                 &b_bit_idx,
                 &(u_star.clone() * &b_star_idx.clone()),
-                &n_preimage_bit_id,
             );
             log_mem("Computed n_preimage_bit");
 
-            n_preimages_paths[idx][bit] = n_preimage_bit_path;
+            n_preimages[idx].push(n_preimage_bit);
 
             let rg = &public_data.rgs[bit];
             let top = lhs.mul_tensor_identity_decompose(rg, 1 + packed_input_size);
@@ -216,16 +214,11 @@ where
             };
             let bottom = pub_key_idx[0].concat_matrix(&pub_key_idx[1..]) - &inserted_poly_gadget;
             let k_target = top.concat_rows(&[&bottom]);
-            let k_preimage_bit_path = sampler_trapdoor.preimage_to_fs(
-                &params,
-                &b_bit_trapdoor_idx,
-                &b_bit_idx,
-                &k_target,
-                &k_preimage_bit_id,
-            );
+            let k_preimage_bit =
+                sampler_trapdoor.preimage(&params, &b_bit_trapdoor_idx, &b_bit_idx, &k_target);
             log_mem("Computed k_preimage_bit");
 
-            k_preimages_paths[idx][bit] = k_preimage_bit_path;
+            k_preimages[idx].push(k_preimage_bit);
         }
 
         b_star_trapdoor_cur = b_star_trapdoor_idx;
@@ -259,14 +252,11 @@ where
     };
     log_mem("Computed final_preimage_target");
 
-    let final_preimage_id = "final_preimage";
-
-    let final_preimage_path = sampler_trapdoor.preimage_to_fs(
+    let final_preimage = sampler_trapdoor.preimage(
         &params,
         &b_star_trapdoor_cur,
         &b_star_cur,
         &final_preimage_target,
-        final_preimage_id,
     );
     log_mem("Sampled final_preimage");
 
@@ -275,10 +265,10 @@ where
         enc_hardcoded_key,
         encodings_init,
         p_init,
-        m_preimages_paths,
-        n_preimages_paths,
-        k_preimages_paths,
-        final_preimage_path,
+        m_preimages,
+        n_preimages,
+        k_preimages,
+        final_preimage,
         #[cfg(feature = "test")]
         s_init: s_init.clone(),
         #[cfg(feature = "test")]
