+++ src/poly/sampler.rs
@@ -1,8 +1,6 @@
-use std::path::PathBuf;
-
 use super::{Poly, PolyMatrix};
 
-#[derive(Debug)]
+#[derive(Debug, Clone, Copy)]
 /// Enum representing different types of distributions for random sampling.
 pub enum DistType {
     /// Distribution over a finite ring, typically samples elements from a ring in a uniform or
@@ -53,23 +51,17 @@ pub trait PolyUniformSampler {
 pub trait PolyTrapdoorSampler {
     type M: PolyMatrix;
     type Trapdoor;
-    type MatrixPtr;
 
     fn trapdoor(
         &self,
         params: &<<Self::M as PolyMatrix>::P as Poly>::Params,
         size: usize,
     ) -> (Self::Trapdoor, Self::M);
-    fn preimage_to_fs(
+    fn preimage(
         &self,
         params: &<<Self::M as PolyMatrix>::P as Poly>::Params,
         trapdoor: &Self::Trapdoor,
         public_matrix: &Self::M,
         target: &Self::M,
-        preimage_id: &str,
-    ) -> Vec<PathBuf>;
-    fn preimage_from_fs(
-        params: &<<Self::M as PolyMatrix>::P as Poly>::Params,
-        preimages_paths: &[PathBuf],
     ) -> Self::M;
 }
